[Improvement] Better error handling, sub-org implementation

+++ server/model/businessLogic/userLogic.js
@@ -1,32 +1,35 @@
-const mongoose = require('mongoose');
 const User = require('../dbModel/userModel');
 const URL = require('../dbModel/urlModel')
-const showError = require('../../utils/middleware').showError;
+const AppError = require('../../utils/appError')
 
 // Get information about all the users
-const getAllUsers = async () => {
+const getAllUsers = async (next, role='user') => {
     try {
-        let userList = await User.find().lean();
+        let userList = await User.find({role: role}).lean();
+        if(!userList)
+            return next(new AppError("Something went wrong Try Again.", 404));
         return userList;
     }
     catch(err){
-        showError(err);
+        next(err);
     }
 };
 
 // Get information about a given user
-const getUserInfo = async (email) => {
+const getUserInfo = async (email, next) => {
     try {
         let userInfo = await User.findOne({email: email}).select('+password');
+        if(!userInfo)
+            return next(new AppError("User not found.", 404));
         return userInfo;
     }
     catch(err){
-        showError(err);
+        next(err);
     }
 };
 
 // Create a new user
-const createNewUser = async (user) => {
+const createNewUser = async (user, next) => {
     try{
         let newUser = new User(
             {
@@ -36,70 +39,81 @@ const createNewUser = async (user) => {
                 'email': user.email
             });
         let newUserInfo = await newUser.save();
+        if(!newUserInfo)
+            return next(new AppError("Save failed. Please try again", 500));
         return newUserInfo;
     }
     catch(err){
-        showError(err);
+        next(err);
     }
 };
 
 // Update a given user info
-const updateUser = async (user) => {
+const updateUser = async (user, next) => {
     try{
         let updatedUserInfo = await User.findByIdAndUpdate(user._id, {
             'name' : user.name,
             'blacklisted': user.blacklisted
         }, {new: true});
+        if(!updatedUserInfo)
+            return next(new AppError("Failed to update. Please try again.", 500));
         return updatedUserInfo;
     }
     catch(err){
-        showError(err);
+        next(err);
     }
 };
 
 //Increment the URL count of a user
-const incrementUserURL = async (id) => {
+const incrementUserURL = async (id, next) => {
     try{
         let updatedUserInfo = await User.findOneAndUpdate(
             { _id: id},
             { $inc: { numberOfURLs: 1 }  },
             {new: true});
+        if(!updatedUserInfo)
+            return next(new AppError("Failed to update counter. Please try again.", 500));
         return updatedUserInfo;
     }
     catch(err){
-        showError(err);
+        next(err);
     }
 };
 
 //Decrement the URL count of a user
-const decrementUserURL = async (id) => {
+const decrementUserURL = async (id, next) => {
     try{
         let updatedUserInfo = await User.findOneAndUpdate(
             { _id: id},
             { $inc: { numberOfURLs: -1 }  },
             {new: true});
-
+        if(!updatedUserInfo)
+            return next(new AppError("Failed to update counter. Please try again.", 500));
         return updatedUserInfo;
     }
     catch(err){
-        showError(err);
+        next(err);
     }
 };
 
 //Delete a given user
-const deleteUser = async (id) => {
+const deleteUser = async (id, next) => {
     try{
-        await URL.deleteMany({userID : id});
-        await User.deleteOne({ _id: id});
+        let deletedURls = await URL.deleteMany({userID : id});
+        if(!deletedURls)
+            return next(new AppError("Failed to delete Users URLs first. Please try deleting the user again.", 500));
+        let deletedUser = await User.deleteOne({ _id: id});
+        if(!deletedUser)
+            return next(new AppError("Failed to delete the user. Please try deleting the user again.", 500));
         console.log(`deleted user with id : ${id}`);
     }
     catch(err){
-        showError(err);
+        next(err);
     }
 }
 
 //Blacklist a user
-const blacklistUser = async (id) => {
+const blacklistUser = async (id, next) => {
     try{
         let updatedUserInfo = await User.findByIdAndUpdate(id, {
             blacklisted: true
@@ -107,15 +121,17 @@ const blacklistUser = async (id) => {
         await URL.updateMany({userID: id}, {
             blacklisted: true
         });
+        if(!updatedUserInfo)
+            return next(new AppError("Failed to update user. Please try again.", 500));
         return updatedUserInfo;
     }
     catch (err) {
-        showError(err);
+        next(err);
     }
 }
 
 // Whitelist a User
-const whitelistUser = async (id) => {
+const whitelistUser = async (id, next) => {
     try{
         let updatedUserInfo = await User.findByIdAndUpdate(id, {
             blacklisted: false
@@ -123,10 +139,12 @@ const whitelistUser = async (id) => {
         await URL.updateMany({userID: id}, {
             blacklisted: false
         });
+        if(!updatedUserInfo)
+            return next(new AppError("Failed to update user. Please try again.", 500));
         return updatedUserInfo;
     }
     catch (err) {
-        showError(err);
+        next(err);
     }
 }
